subtract, ADMIN cannot basket see and buy anything, fix bugs

+++ pizza_projekt/src/main/java/whz/pti/eva/pizza_projekt/customer/domain/DTO/TransferDTO.java
@@ -7,12 +7,12 @@ import java.io.Serializable;
 public class TransferDTO implements Serializable{
 
     @Getter String to;
-    @Getter int amount;
+    @Getter double amount;
 
     public TransferDTO() {
     }
 
-    public TransferDTO(String to, int amount) {
+    public TransferDTO(String to, double amount) {
         this.to = to;
         this.amount = amount;
     }
